backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-multipart
lxml
requests
scikit-learn
numpy
scipy
python-dotenv
reportlab>=4.0.0
httpx
langchain
langchain-openai
langchain-tavily